Extract Event clas to own file
