+++ filenames.gypi
@@ -63,6 +63,7 @@
       'lib/renderer/api/remote.js',
       'lib/renderer/api/screen.js',
       'lib/renderer/api/web-frame.js',
+      'lib/renderer/extensions/event.js',
       'lib/renderer/extensions/i18n.js',
       'lib/renderer/extensions/storage.js',
       'lib/renderer/extensions/web-navigation.js',
