+++ lib/renderer/chrome-api.js
@@ -1,30 +1,8 @@
 const {ipcRenderer} = require('electron')
+const Event = require('./extensions/event')
 const url = require('url')
 
 let nextId = 0
-
-class Event {
-  constructor () {
-    this.listeners = []
-  }
-
-  addListener (callback) {
-    this.listeners.push(callback)
-  }
-
-  removeListener (callback) {
-    const index = this.listeners.indexOf(callback)
-    if (index !== -1) {
-      this.listeners.splice(index, 1)
-    }
-  }
-
-  emit (...args) {
-    for (const listener of this.listeners) {
-      listener(...args)
-    }
-  }
-}
 
 class Tab {
   constructor (tabId) {
