+++ lib/renderer/extensions/web-navigation.js
@@ -1,19 +1,5 @@
-exports.onBeforeNavigate = {
-  addListener () {
+const Event = require('./event')
 
-  },
+exports.onBeforeNavigate = new Event()
 
-  removeListener () {
-
-  }
-}
-
-exports.onCompleted = {
-  addListener () {
-
-  },
-
-  removeListener () {
-
-  }
-}
+exports.onCompleted = new Event()
